+++ app/app/src/main/java/de/stustapay/stustapay/ui/user/UserLoginView.kt
@@ -12,11 +12,11 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.compose.ui.window.Dialog
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import com.ionspin.kotlin.bignum.integer.toBigInteger
 import de.stustapay.stustapay.R
 import de.stustapay.stustapay.model.UserRolesState
-import de.stustapay.stustapay.model.UserTag
+import de.stustapay.api.models.UserTag
 import de.stustapay.stustapay.ui.chipscan.NfcScanDialog
 import de.stustapay.stustapay.ui.chipscan.rememberNfcScanDialogState
 import de.stustapay.stustapay.ui.theme.errorButtonColors
@@ -109,7 +109,7 @@ fun UserLoginView(
                         LaunchedEffect(Unit) {
                             scope.launch {
                                 roleSelection = RoleSelectionState.Closed
-                                viewModel.login(roleSelectionV.tag, userRolesV.roles[0].id)
+                                viewModel.login(roleSelectionV.tag, userRolesV.roles[0].id.intValue())
                             }
                         }
                     } else {
